add maven failsafe plugin for running integration tests
