+++ pom.xml
@@ -16,6 +16,7 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <maven.compiler.source>1.8</maven.compiler.source>
         <maven.compiler.target>1.8</maven.compiler.target>
+        <skipITs>true</skipITs>
     </properties>
     <dependencies>
         <!-- https://mvnrepository.com/artifact/org.codehaus.mojo/build-helper-maven-plugin -->
@@ -180,13 +181,27 @@
                     <excludes>
                         <exclude>**/LargeBatchTest.java</exclude>
                         <exclude>**/MultiThreadedVolumeTest.java</exclude>
-                        <exclude>**/ReconciliationTest.java</exclude> 
-                        <exclude>**/ByteBufferWithMixedEventsTest.java</exclude> 
                         <exclude>**/OncePerSecondLongevityTest.java</exclude>
-                        <exclude>**/AWSSourcetypeTest.java</exclude>
                     </excludes>
                 </configuration>
-            </plugin>            
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-failsafe-plugin</artifactId>
+                <version>2.20.1</version>
+                <configuration>
+                    <argLine>-Xmx16G</argLine><!-- set heap memory for mvn junit tests -->
+                    <skipITs>${skipITs}</skipITs>
+                </configuration>
+                <executions>
+                    <execution>
+                        <goals>
+                            <goal>integration-test</goal>
+                            <goal>verify</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
         </plugins>
     </build>  
     <licenses>
